email_assistant/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.54.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.2.0
asyncpg>=0.29.0
neo4j>=5.20.0
pdfminer.six>=20231228
openpyxl>=3.1.0
mammoth>=1.8.0
python-pptx>=1.0.2
beautifulsoup4>=4.12.0
markdownify>=0.12.0
markitdown>=0.0.1a3
python-dateutil>=2.9.0
